+++ src/collections/blog/hooks/revalidate-post.ts
@@ -1,6 +1,4 @@
-'use server'
-
-import { revalidatePath, revalidateTag } from 'next/cache'
+import { revalidatePath, updateTag } from 'next/cache'
 
 import type { CollectionAfterChangeHook, CollectionAfterDeleteHook } from 'payload'
 import type { Blog } from '@/payload-types'
@@ -17,8 +15,8 @@ export const revalidatePost: CollectionAfterChangeHook<Blog> = async ({
       payload.logger.info(`Revalidating post at path: ${path}`)
 
       revalidatePath(path)
-      revalidateTag(`blog-${doc.slug}`, 'max')
-      revalidateTag('blog-sitemap', 'max')
+      updateTag(`blog-${doc.slug}`)
+      updateTag('blog-sitemap')
     }
 
     // If the post was previously published, we need to revalidate the old path
@@ -28,8 +26,8 @@ export const revalidatePost: CollectionAfterChangeHook<Blog> = async ({
       payload.logger.info(`Revalidating old post at path: ${oldPath}`)
 
       revalidatePath(oldPath)
-      revalidateTag(`blog-${previousDoc.slug}`, 'max')
-      revalidateTag('blog-sitemap', 'max')
+      updateTag(`blog-${previousDoc.slug}`)
+      updateTag('blog-sitemap')
     }
   }
   return doc
@@ -43,8 +41,8 @@ export const revalidateDelete: CollectionAfterDeleteHook<Blog> = async ({
     const path = `/blog/${doc?.slug}`
 
     revalidatePath(path)
-    revalidateTag(`blog-${doc.slug}`, 'max')
-    revalidateTag('blog-sitemap', 'max')
+    updateTag(`blog-${doc.slug}`)
+    updateTag('blog-sitemap')
   }
 
   return doc