+++ src/collections/global/hooks/revalidate-global.ts
@@ -1,4 +1,6 @@
-import { revalidateTag } from 'next/cache'
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+import { updateTag } from 'next/cache'
 
 import type { PayloadRequest, RequestContext, SanitizedGlobalConfig } from 'payload'
 
@@ -20,7 +22,7 @@ export const revalidateGlobal = (
   if (!context.disableRevalidate) {
     payload.logger.info(`✨ Revalidating ${tag}`)
 
-    revalidateTag(tag, 'max')
+    updateTag(tag)
   }
 
   return doc