+++ src/lib/payload/get-globals.ts
@@ -1,4 +1,4 @@
-import { cacheTag } from 'next/cache'
+import { cacheLife, cacheTag } from 'next/cache'
 
 import configPromise from '@payload-config'
 import { getPayload } from 'payload'
@@ -8,6 +8,7 @@ type Global = keyof Config['globals']
 
 export async function getGlobal(slug: Global, depth = 0) {
   'use cache'
+  cacheLife('hours')
   cacheTag(slug)
 
   const payload = await getPayload({ config: configPromise })